feat: 이미지 관련 기능 수정 및 추가

+++ backend/src/main/java/com/project/vegan/domain/post/service/PostService.java
@@ -128,7 +128,10 @@ public class PostService {
         }
 
         hashTagRepository.deleteAll(hashTagRepository.findByPost(post));
-        postUploadFileRepository.deleteAll(postUploadFileRepository.findByPost(post));
+
+        if(postUploadFileRepository.existsByPost(post)){
+            postUploadFileRepository.deleteAll(postUploadFileRepository.findByPost(post));
+        }
 
         postRepository.delete(post);
     }
